requirements.txt
flask
werkzeug
numpy
flask-restplus
pytest
pytest-runner
pytest-cov
Flask-JWT
pytest-flask
Flask-Migrate
Flask-SQLAlchemy
